requirements.txt
requests==2.25.1
Pillow==8.1.0
discord.py==1.5.1